test: separate the mock data for events
