+++ apps/calendar/stories/dayView.stories.tsx
@@ -5,8 +5,9 @@ import { Story } from '@storybook/preact';
 import { Day } from '@src/components/view/day';
 import EventModel from '@src/model/eventModel';
 import TZDate from '@src/time/date';
-import { addDate, setTimeStrToDate } from '@src/time/datetime';
+import { addDate } from '@src/time/datetime';
 
+import { mockDayViewEvents } from '@stories/mocks/mockDayViewEvents';
 import { ProviderWrapper } from '@stories/util/providerWrapper';
 import { createRandomEventModelsForMonth, createRandomEvents } from '@stories/util/randomEvents';
 
@@ -22,58 +23,7 @@ function createTimeGridEvents() {
   return createRandomEvents('week', start, end).map((event) => EventModel.create(event));
 }
 
-function createDayEvents() {
-  const today = new TZDate();
-  const yesterday = addDate(today, -1);
-  const tomorrow = addDate(today, 1);
-  const events: EventModelData[] = [
-    {
-      id: '1',
-      calendarId: 'cal1',
-      title: 'event1',
-      category: 'allday',
-      isAllday: true,
-      start: yesterday,
-      end: today,
-    },
-    {
-      id: '2',
-      calendarId: 'cal1',
-      title: 'event2',
-      category: 'allday',
-      isAllday: true,
-      start: today,
-      end: today,
-    },
-    {
-      id: '3',
-      calendarId: 'cal1',
-      title: 'event3',
-      category: 'allday',
-      isAllday: true,
-      start: today,
-      end: tomorrow,
-    },
-    {
-      id: '4',
-      calendarId: 'cal1',
-      title: 'long time event',
-      category: 'time',
-      isAllday: false,
-      start: setTimeStrToDate(yesterday, '10:00'),
-      end: setTimeStrToDate(today, '06:00'),
-    },
-    {
-      id: '5',
-      calendarId: 'cal1',
-      title: 'short time event',
-      category: 'time',
-      isAllday: false,
-      start: setTimeStrToDate(today, '04:00'),
-      end: setTimeStrToDate(today, '06:00'),
-    },
-  ];
-
+function createDayEvents(events: EventModelData[]) {
   return events.map((event) => EventModel.create(event));
 }
 
@@ -94,5 +44,5 @@ randomEvents.args = {
 export const FixedEvents = Template.bind({});
 FixedEvents.args = {
   options: { useCreationPopup: true, useDetailPopup: true },
-  events: createDayEvents(),
+  events: createDayEvents(mockDayViewEvents),
 };
