+++ apps/calendar/stories/monthView.stories.tsx
@@ -1,13 +1,11 @@
 import { h } from 'preact';
 
 import { Story } from '@storybook/preact';
-import range from 'tui-code-snippet/array/range';
 
 import { Month } from '@src/components/view/month';
 import EventModel from '@src/model/eventModel';
-import TZDate from '@src/time/date';
-import { addDate } from '@src/time/datetime';
 
+import { mockMonthViewEvents } from '@stories/mocks/mockMonthViewEvents';
 import { ProviderWrapper } from '@stories/util/providerWrapper';
 import { createRandomEventModelsForMonth } from '@stories/util/randomEvents';
 
@@ -15,54 +13,7 @@ import { EventModelData } from '@t/events';
 
 export default { title: 'Views/MonthView', component: Month };
 
-function createMonthEvents() {
-  const DAYS_OF_WEEK = 7;
-  const today = new TZDate();
-  const thisSunday = addDate(today, -today.getDay());
-  const sundayDate = thisSunday.getDate();
-  const sundayMonth = thisSunday.getMonth();
-  const todayMonth = today.getMonth();
-  const weekCount =
-    sundayMonth !== todayMonth
-      ? -1
-      : Math.floor(
-          sundayDate % DAYS_OF_WEEK ? sundayDate / DAYS_OF_WEEK : (sundayDate - 1) / DAYS_OF_WEEK
-        );
-  const firstSunday = addDate(thisSunday, -weekCount * DAYS_OF_WEEK);
-  const firstTuesday = addDate(firstSunday, 2);
-  const secondTuesday = addDate(firstTuesday, DAYS_OF_WEEK);
-  const secondThursday = addDate(secondTuesday, 2);
-  const thirdThursday = addDate(secondThursday, DAYS_OF_WEEK);
-  const thirdSaturday = addDate(thirdThursday, 2);
-  const events: EventModelData[] = [
-    {
-      title: 'event1',
-      start: firstSunday,
-      end: secondTuesday,
-      id: '0',
-    },
-    {
-      title: 'event2',
-      start: secondTuesday,
-      end: secondThursday,
-      id: '1',
-    },
-    {
-      title: 'event3',
-      start: thirdThursday,
-      end: thirdSaturday,
-      id: '2',
-    },
-  ];
-  range(10).forEach((i) => {
-    events.push({
-      title: `event2-${i}`,
-      start: secondTuesday,
-      end: secondThursday,
-      id: `${i}${i}`,
-    });
-  });
-
+function createMonthEvents(events: EventModelData[]) {
   return events.map((event) => EventModel.create(event));
 }
 
@@ -112,5 +63,5 @@ randomEvents.args = {
 export const FixedEvents = Template.bind({});
 FixedEvents.args = {
   options: { useCreationPopup: true, useDetailPopup: true },
-  events: createMonthEvents(),
+  events: createMonthEvents(mockMonthViewEvents),
 };
